fix: pass bounded signature ts test
