+++ src/js/wasm_module.js
@@ -143,12 +143,16 @@ module.exports.blsVerifyProofMulti = async (request) => {
 
 module.exports.boundedBlsSignatureRequest = async (request) => {
   await initialize();
-  return await throwErrorOnRejectedPromise(wasm.boundedBlsSignatureRequest(request));
+  return await throwErrorOnRejectedPromise(
+    wasm.boundedBlsSignatureRequest(request)
+  );
 };
 
 module.exports.verifyBoundedBlsSignatureRequest = async (request) => {
   await initialize();
-  return await throwErrorOnRejectedPromise(wasm.verifyBoundedBlsSignatureRequest(request));
+  return await throwErrorOnRejectedPromise(
+    wasm.verifyBoundedBlsSignatureRequest(request)
+  );
 };
 
 module.exports.boundedBlsSign = async (request) => {
@@ -158,5 +162,7 @@ module.exports.boundedBlsSign = async (request) => {
 
 module.exports.unblindBoundedBlsSignature = async (request) => {
   await initialize();
-  return await throwErrorOnRejectedPromise(wasm.unblindBoundedBlsSignature(request));
+  return await throwErrorOnRejectedPromise(
+    wasm.unblindBoundedBlsSignature(request)
+  );
 };
